+++ FROST-Server.SQL/src/main/java/de/fraunhofer/iosb/ilt/frostserver/persistence/postgres/factories/DatastreamFactory.java
@@ -144,11 +144,6 @@ public class DatastreamFactory<I extends SimpleExpression<J> & Path<J>, J> imple
         insert.set(qd.unitSymbol, ds.getUnitOfMeasurement().getSymbol());
         insert.set(qd.properties, EntityFactories.objectToJson(ds.getProperties()));
 
-        insert.set(qd.phenomenonTimeStart, new Timestamp(PostgresPersistenceManager.DATETIME_MAX.getMillis()));
-        insert.set(qd.phenomenonTimeEnd, new Timestamp(PostgresPersistenceManager.DATETIME_MIN.getMillis()));
-        insert.set(qd.resultTimeStart, new Timestamp(PostgresPersistenceManager.DATETIME_MAX.getMillis()));
-        insert.set(qd.resultTimeEnd, new Timestamp(PostgresPersistenceManager.DATETIME_MIN.getMillis()));
-
         insert.set(qd.getObsPropertyId(), (J) op.getId().getValue());
         insert.set(qd.getSensorId(), (J) s.getId().getValue());
         insert.set(qd.getThingId(), (J) t.getId().getValue());