+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Version 1.10 is not released yet.
 **Bugfixes**
 * Fixed #110, paths with multiple ids (v1.0/Things(1)/Datastreams(2)/...) not
   always properly validated.
+* Fixed #???, filtering on Datastream/(phenomenon|result)Time returns Datastreams
+  that have no such Time.
 
 
 # Release Version 1.9